The LED's Annunciate Version 2.2.  2 long FWD LED flashes followed by 2 short RIGHT LED flashes.
When Delay Pot is fully CCW (or near), the USER PORT Switch drives the
MODE/RESET signal to the wheelchair electronics. As long as the USER
PORT switch is closed, the MODE/RESET signal is active (asserted).

In Bluetooth mode, the Reverse Pad activation sends a Reverse signal to
the Bluetooth and, ultimately, to the Bluetooth connected device, i.e.
Mouse Mover.